change mysql.sql int to bigint

+++ db/mysql/mysql.sql
@@ -3,19 +3,19 @@ create database competence default character set utf8 collate utf8_general_ci;
 
 use competence;
 create table group_info(
-	id integer(10) primary key not null,
+	id bigint(10) primary key not null,
 	group_num varchar(200),
 	group_name varchar(255),
 	description varchar(255),	
 	use_flag tinyint(1),
-	create_user integer(10),
+	create_user bigint(10),
 	create_time timestamp,
-	update_user integer(10),
+	update_user bigint(10),
 	update_time timestamp
 )engine=innodb default charset=utf8;
 
 create table user(
-	id integer(10) primary key not null,
+	id bigint(10) primary key not null,
 	name varchar(200),
 	login_name varchar(200),
 	password varchar(255),
@@ -26,88 +26,88 @@ create table user(
 	phone varchar(20),
 	description varchar(255),
 	use_flag tinyint(1),
-	group_id integer(10),
-	create_user integer(10),
+	group_id bigint(10),
+	create_user bigint(10),
 	create_time timestamp,
-	update_user integer(10),
+	update_user bigint(10),
 	update_time timestamp,
 foreign key(group_id) references group_info(id) on delete cascade on update cascade
 )engine=innodb default charset=utf8;
 
 create table group_user(
-	id integer(10) primary key not null,
+	id bigint(10) primary key not null,
 	group_num varchar(200),
 	login_name varchar(200),
-	user_id integer(10),
-	group_id integer(10),
-	create_user integer(10),
+	user_id bigint(10),
+	group_id bigint(10),
+	create_user bigint(10),
 	create_time timestamp,
-	update_user integer(10),
+	update_user bigint(10),
 	update_time timestamp
 )engine=innodb default charset=utf8;
 
 create table role(
-	id integer(10) primary key not null,
+	id bigint(10) primary key not null,
 	role_name varchar(255),
 	role_num varchar(200),
 	description varchar(255),
 	use_flag tinyint(1),
-	create_user integer(10),
+	create_user bigint(10),
 	create_time timestamp,
-	update_user integer(10),
+	update_user bigint(10),
 	update_time timestamp
 )engine=innodb default charset=utf8;
 
 create table user_role(
-	id integer(10) primary key not null,
+	id bigint(10) primary key not null,
 	login_name varchar(200),
 	role_num varchar(200),
 	description varchar(255),
-	user_id integer(10),
-	role_id integer(10),
-	create_user integer(10),
+	user_id bigint(10),
+	role_id bigint(10),
+	create_user bigint(10),
 	create_time timestamp,
-	update_user integer(10),
+	update_user bigint(10),
 	update_time timestamp,
 foreign key(user_id) references user(id) on delete cascade on update cascade,
 foreign key(role_id) references role(id) on delete cascade on update cascade
 )engine=innodb default charset=utf8;
 
 create table menu(
-	id integer(10) primary key not null,
+	id bigint(10) primary key not null,
 	menu_num varchar(200),
 	menu_name varchar(100),
-	parent_menu integer(10),
+	parent_menu bigint(10),
 	use_flag tinyint(1),
 	description varchar(255),
-	create_user integer(10),
+	create_user bigint(10),
 	create_time timestamp,
-	update_user integer(10),
+	update_user bigint(10),
 	update_time timestamp
 )engine=innodb default charset=utf8;
 
 create table menu_project(
-	id integer(10) primary key not null,
+	id bigint(10) primary key not null,
 	project varchar(200),
 	project_name varchar(100),
-	menu_id integer(10),
-	create_user integer(10),
+	menu_id bigint(10),
+	create_user bigint(10),
 	create_time timestamp,
-	update_user integer(10),
+	update_user bigint(10),
 	update_time timestamp,
 foreign key(menu_id) references menu(id) on delete cascade on update cascade
 )engine=innodb default charset=utf8;
 
 
 create table user_menu(
-	id integer(10) primary key not null,
+	id bigint(10) primary key not null,
 	description varchar(255),
-	user_id integer(10),
-	menu_id integer(10),
-	project_id integer(10),
-	create_user integer(10),
+	user_id bigint(10),
+	menu_id bigint(10),
+	project_id bigint(10),
+	create_user bigint(10),
 	create_time timestamp,
-	update_user integer(10),
+	update_user bigint(10),
 	update_time timestamp,
 foreign key(user_id) references user(id) on delete cascade on update cascade,
 foreign key(menu_id) references menu(id) on delete cascade on update cascade,
@@ -115,14 +115,14 @@ foreign key(project_id) references menu_project(id) on delete cascade on update 
 )engine=innodb default charset=utf8;
 
 create table role_menu(
-	id integer(10) primary key not null,
+	id bigint(10) primary key not null,
 	description varchar(255),
-	role_id integer(10),
-	menu_id integer(10),
-	project_id integer(10),
-	create_user integer(10),
+	role_id bigint(10),
+	menu_id bigint(10),
+	project_id bigint(10),
+	create_user bigint(10),
 	create_time timestamp,
-	update_user integer(10),
+	update_user bigint(10),
 	update_time timestamp,
 foreign key(role_id) references role(id) on delete cascade on update cascade,
 foreign key(menu_id) references menu(id) on delete cascade on update cascade,
@@ -130,25 +130,25 @@ foreign key(project_id) references menu_project(id) on delete cascade on update 
 )engine=innodb default charset=utf8;
 
 create table op_info(
-	id integer(10) primary key not null,
+	id bigint(10) primary key not null,
 	op_num varchar(200),
 	op_name varchar(255),
 	description varchar(255),
-	create_user integer(10),
+	create_user bigint(10),
 	create_time timestamp,
-	update_user integer(10),
+	update_user bigint(10),
 	update_time timestamp
 )engine=innodb default charset=utf8;
 
 create table menu_op(
-	id integer(10) primary key not null,
+	id bigint(10) primary key not null,
 	use_flag tinyint(1),
-	menu_id integer(10),
-	operator_id integer(10),
-	project_id integer(10),
-	create_user integer(10),
+	menu_id bigint(10),
+	operator_id bigint(10),
+	project_id bigint(10),
+	create_user bigint(10),
 	create_time timestamp,
-	update_user integer(10),
+	update_user bigint(10),
 	update_time timestamp,
 foreign key(menu_id) references menu(id) on delete cascade on update cascade,
 foreign key(operator_id) references op_info(id) on delete cascade on update cascade,
